Move CachedSettings to setting file
